Rename client/service.rs -> client/connection.rs
