+++ src/client/mod.rs
@@ -1,7 +1,7 @@
 mod background;
 mod new_service;
-mod service;
+mod connection;
 
 pub use self::background::Background;
 pub use self::new_service::{Client, ConnectFuture, ConnectError};
-pub use self::service::{Connection, ResponseFuture, Error};
+pub use self::connection::{Connection, ResponseFuture, Error};
